+++ FuzzySharp/SimilarityRatio/Scorer/StrategySensitive/Simple/ApproximateSubstringRatioScorer.cs
@@ -1,0 +1,9 @@
+using Raffinert.FuzzySharp.SimilarityRatio.Strategy;
+
+namespace Raffinert.FuzzySharp.SimilarityRatio.Scorer.StrategySensitive;
+
+public sealed class ApproximateSubstringRatioScorer : SimpleRatioScorerBase
+{
+    protected override FuzzySharp.Scorer Scorer =>
+        ApproximateSubstringRatioStrategy.Calculate;
+}